Add SetNZCount to OmmonMatrix to support setting number of Non-zero values externally.
Change the Resize function in sparse matrices to make it clear that the numNZ is used to reserve memory only so that we can call Resize repeatedly without affecting the actual number of non-zero values.
Change the cpu sparse matrix's Resize to support keeping existing values when memory is reallocated.
Change SetValue function in cpu sparsematrix to support automatic resizing.