Implement minibatch prefetching
This change implements a read-ahead (prefetch) of one minibatch on a separate thread to ensure that main thread always has enough work. This is done through coarse-level parallelism, where GetMinibatch is called from a separate thread and result is cached for main thread to consume. The synchronization is a simple producer/consumer model with a single mutex and a conditional variable.

This change also modifies how we measure time it takes to read and compute a minibatch worth of data.