changed feature packing flag names from utteranceXYZ and sentence_XYZ to SequenceXYZ to make them consistent and be understood for both speech and text people.
Added flag to indicate the condition of no feature (with assumption that there is also no label)  to support efficient bi-directional model training.
Updated the windows version of HTKMLFReader, LM and LU Sequence readers.
Caution: this change will not allow for using past and future label information in the recurrent networks. The ultimate solution is to have a flag for each feature read-in. However, this will require huge change and we will do this after reader refactorization to reduce effort.