Add a configure script for setting build parameters
Add a configure script for initializing build parameters, either
for in or out of source builds. The script generates a Config.make
in the build directory, and, for out of source builds, a trampoline
Makefile.

Make the build-and-test script to do an out of source build.

Add Config.make to .gitignore, as well as emacs temporary file patterns.