Replace .docx documentation files with Markdown
Went through and makes a GitHub flavor Markdown file for each .docx file
in the Documentation/Documents folder. There will inevitably need to be
changes and upgrades to this, but I think it's a good first step.

As requested, this also removes the .docx files from
Documentation/Documents as well.