Fix CntkEval.dll Related Issues
Consumers of the CntkEval interface can pass in feature values
without label values. Most of the code was okay with this, but
the SimpleOutputWriter insisted on initializing both features
and labels. This caused GetMinibatchIntoNetwork function in
DataReaderHelpers.h to fail, when the layout of the label node
was incorrect.

I've changed the SimpleOutputWriter to only use input nodes
that are ancestors of the desired output nodes. This fixes the
issue for CntkEval usage.