Fix PerDimMeanVarNormalizationNode to Work With Legacy Models
Recent changes in the default tensor shape for the inputs has broken
CNTK trying to load and evaluate older models that contain this node.
This fix checks the sample layouts of the data, and the mean and variance,
and infers the shape of the latter from the former.

This code will copy the shape of Input(0) (source) to Input(1) and Input(2)
(target) if:
  1. The source is a 3-tensor with shape 1x1xM
  2. The target is a vector (i.e., a 2-tensor with shape Nx1)
  3. Both targets have the same number of elements
  4. The number of elements in the target (N) is the same as the number of
     elelemtns in the source (M)