This commit adds the CIFAR-10 Examples as end to end tests Each test includes its run-test and testcases.yaml file and baselines.
This commit fixes the Example for Image/Miscellaneous/CIFAR-10 to work with data downloaded from the scripts without having to manipulate it first.
This commit also makes EndToEndTests out of the CIFAR-10 Example(s)
The EndToEnd tests copy the network shared data to the local folder for prior to running (if the data doesn't exist locally first)
It also fixes the access to external data for CIFAR-10 when running inside cygwin
It also modifies the CIFAR cofigurations to initialize on CPU.
It also sets the maxEpochs to reasonable values
It also adds the CIFAR README.md file to the solution.
This commit adds a note to the 01_Conv configuration file in CIFAR-10 EndToEndTests examples, indicating why minibatchMode="full" is needed.
It also updates the execution times in the yml files for the CIFAR-10 E2E Tests