Move workspace and reserve matricies from CuDnnRNN to RNNNode.
These matricies need to be managed together with the rest of the
related matricies, to ensure they are on the same device. It is
easiest to do this, by exposing them and bringing them into the
node's class. As a happy result, there is no longer gobs of GPU
memory being silently eaten by the library - it is explicit at
the node level.