Fix Memshare Bug in RNNNode
The RNNNode doesn't technically use the input data during backprop, but the code
did query the matrix size. When the matrix was shared, it is possible that this
size gets changed by someone else before it is read out.

The fix is to query the node directly for the number of elements in each piece of
data.

I've also marked the parameters input as not shared, in case someone does something
weird on the input side in the future.

Finally, the check that Input(1) is data has been moved from forwardprop to the
validation code.