Fix timestamp bump
After ResetEvalTimeStamps function is called, all nodes in the network
will have the same timestamp equal to global timestamp counter. The
first BumpEvalTimeStamp function that is called after that will leave
the timestamp of the node unchanged and will just increment global
counter. This causes the problem when one wants to repeat forward pass
for the part of the network, by bumping timestamps of the nodes from
which the forward pass should be performed. E.g. from accumulator nodes
to evaluation nodes after aggregation of accumulator values is done at
epoch end.
This change fixes described problem.