BS.Constants.Range(dim) as 1D constant tensor
The `Range(dim)` is intended to support the creation
of useful masks which are handy when dealing, for example,
with missing data.

The function `Range(dim)` generates a 1D constant tensor
that contains integers numbered from 0 .. (dim -1).

While the implementation is not complicated, it's a
non-trivial use of the primitives provided by BrainScript,
hence the justification for an inclusion in the core lib.

Microsoft/CNTK#2121