BrainScript: LogGamma(x) as a special function
The Gamma function plays an important role in the definition
of many probability distributions. In particular, the log of
the Gamma function is a key ingredient to train a network
over negative binomial distributions, which, under specific
situations, represent a superior alternative to normal
distributions.

The `LogGamma(x)` function is approximated based on the method
of the continued fractions. The hard-coded values have been
obtained from _NIST Handbook of Mathematical Functions_ (2010).