.NET binary writer, optimized for columnar I/O
A .NET writer for the generic binary format of CNTK. The implementation is
optimized for high-performance columnar-to-columnar reformatting.

The implementation features online behavior that ensures that chunks are flushed
out to the output stream as the input data get written. This behavior ensures
that the written operates with constant memory, capped by the size of a chunk.

The API also provides method overloads intended to support object pooling, a
typical .NET trick to increase throughput while processing large amount of data.

The C# project is kept as stand-alone in order to avoid interfering with the
primary C++ solution.

See also
microsoft/CNTK#2058