Remove unnecessary sys/io.h
Machines with aarch64 architecure such as NVIDIA Jetson TX1 does
not have sys/io.h. Without this commit, building
libCntk.Core-2.5.so fails on a aarch64 machine.

It should be safe to completely remove the inclusion of sys/io.h,
since it seems that functions in the header are not used by CNTK.