Fix overlapping memcpy reported by valgrind in SmallVector
Valgrind sometimes reports "Source and destination overlap in memcpy" on
the overloaded = operator of SmallVector class. Whenever it is reported,
the src address and the dest address of memcpy is actually identical.

The solution in this commit is to check the src and dest address first,
and do memcpy only if they are not identical. This should be enough to
handle the issue, since the only case they overlap is when they are
identical, provided that the m_data is a member of a class.